Rename PolicyProvider and file to AppStateProvider
