+++ components/occurrences/OccurrencePreview.js
@@ -19,7 +19,7 @@ import PropTypes from "prop-types";
 import styles from "styles/modules/Occurrences.module.scss";
 import Icon from "components/Icon";
 import { ICON_NAMES } from "utils/icon-utils";
-import { usePolicies } from "providers/policies";
+import { usePolicies } from "providers/appState";
 import { policyActions } from "reducers/policies";
 
 const OccurrencePreview = ({
