+++ components/playground/ResourceSelectionDrawer.js
@@ -23,7 +23,7 @@ import Icon from "components/Icon";
 import { ICON_NAMES } from "utils/icon-utils";
 import ResourceSearchAndResults from "./ResourceSearchAndResults";
 import ResourceVersionSearchAndResults from "./ResourceVersionSearchAndResults";
-import { usePolicies } from "providers/policies";
+import { usePolicies } from "providers/appState";
 import { policyActions } from "reducers/policies";
 
 const RESOURCE = "Resource";
