+++ components/policies/PolicyBreadcrumbs.js
@@ -17,7 +17,7 @@
 import React from "react";
 import styles from "styles/modules/Search.module.scss";
 import Link from "next/link";
-import { usePolicies } from "providers/policies";
+import { usePolicies } from "providers/appState";
 import { SEARCH_ALL } from "utils/constants";
 
 const getSearchTermText = (searchTerm) => {
