+++ components/policies/PolicyForm.js
@@ -25,7 +25,7 @@ import { schema } from "schemas/policy-form";
 import { useFormValidation } from "hooks/useFormValidation";
 import { showError, showSuccess } from "utils/toast-utils";
 import PolicyValidationResult from "components/policies/PolicyValidationResult";
-import { usePolicies } from "providers/policies";
+import { usePolicies } from "providers/appState";
 import { policyActions } from "reducers/policies";
 import Modal from "components/Modal";
 import PageHeader from "components/layout/PageHeader";
