+++ components/policies/PolicySearchBar.js
@@ -17,7 +17,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import SearchBar from "components/shared/SearchBar";
-import { usePolicies } from "providers/policies";
+import { usePolicies } from "providers/appState";
 import { policyActions } from "reducers/policies";
 import { SEARCH_ALL } from "utils/constants";
