+++ components/policies/PolicySearchResult.js
@@ -18,7 +18,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import Button from "components/Button";
 import { useRouter } from "next/router";
-import { usePolicies } from "providers/policies";
+import { usePolicies } from "providers/appState";
 import { policyActions } from "reducers/policies";
 import styles from "styles/modules/Search.module.scss";
 import LabelWithValue from "components/LabelWithValue";
