+++ components/resources/ChangeVersionDrawer.js
@@ -33,7 +33,7 @@ import {
 import { DATE_TIME_FORMAT, SEARCH_ALL } from "utils/constants";
 import dayjs from "dayjs";
 import ResourceVersionSearchBar from "components/resources/ResourceVersionSearchBar";
-import { usePolicies } from "providers/policies";
+import { usePolicies } from "providers/appState";
 import { policyActions } from "reducers/policies";
 
 const ChangeVersionDrawer = (props) => {
