+++ pages/index.js
@@ -19,7 +19,7 @@ import styles from "styles/modules/Home.module.scss";
 import ResourceSearchBar from "components/resources/ResourceSearchBar";
 import PolicySearchBar from "components/policies/PolicySearchBar";
 import { useTheme } from "providers/theme";
-import { usePolicies } from "providers/policies";
+import { usePolicies } from "providers/appState";
 import { policyActions } from "reducers/policies";
 import Link from "next/link";
 import { useRouter } from "next/router";
