+++ pages/policy-groups.js
@@ -22,7 +22,7 @@ import PageHeader from "components/layout/PageHeader";
 import { usePaginatedFetch } from "hooks/usePaginatedFetch";
 import Loading from "components/Loading";
 import styles from "styles/modules/PolicyGroupDashboard.module.scss";
-import { usePolicies } from "providers/policies";
+import { usePolicies } from "providers/appState";
 import { policyActions } from "reducers/policies";
 
 const PolicyGroups = () => {
